requirements.txt
asn1crypto==0.24.0
boto3==1.7.45
botocore==1.10.45
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
docutils==0.14
idna==2.7
jmespath==0.9.3
mkl-fft==1.0.2
mkl-random==1.0.1
numpy==1.14.5
pandas==0.23.3
psycopg2==2.7.5
pycparser==2.18
pyOpenSSL==18.0.0
PySocks==1.6.8
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.8
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
